Refactor more integration-y tests out.
